requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
drf-writable-nested==0.6.3
factory-boy==3.2.1
Faker==11.1.0
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
